refactor api environment files
