+++ src/app/api/session/http/http-session-api.service.ts
@@ -17,7 +17,7 @@ import { Person } from 'src/app/models/entities/Person';
 import { User } from 'src/app/models/entities/User';
 import { HttpService } from 'src/app/shared/http.abstract-service';
 import { SessionApiIService } from '../session-api.iservice';
-import { sessionStorageTokenItemName, authorizationHeaderName, baseURI } from 'src/environments/auth.environment';
+import { sessionStorageTokenItemName, authorizationHeaderName, baseURI } from 'src/environments/auth-api.environment';
 
 @Injectable()
 export class HttpSessionApiService
